Extract reapers into separate classes
The one class was having too many responsibilities
