+++ lib/sidekiq_unique_jobs/constants.rb
@@ -30,6 +30,7 @@ module SidekiqUniqueJobs
   ON_CLIENT_CONFLICT    ||= "on_client_conflict"
   ON_CONFLICT           ||= "on_conflict"
   ON_SERVER_CONFLICT    ||= "on_server_conflict"
+  PROCESSES             ||= "processes"
   QUEUE                 ||= "queue"
   RETRY                 ||= "retry"
   SCHEDULE              ||= "schedule"
