+++ spec/sidekiq_unique_jobs/orphans/reaper_spec.rb
@@ -29,50 +29,6 @@ RSpec.describe SidekiqUniqueJobs::Orphans::Reaper do
     SidekiqUniqueJobs.enable!
   end
 
-  describe "#orphans" do
-    subject(:orphans) { service.orphans }
-
-    context "when scheduled" do
-      let(:item) { raw_item.merge("at" => Time.now.to_f + 3_600) }
-
-      context "without scheduled job" do
-        it { is_expected.to match_array([digest]) }
-      end
-
-      context "with scheduled job" do
-        before { push_item(item) }
-
-        it { is_expected.to match_array([]) }
-      end
-    end
-
-    context "when retried" do
-      let(:item) { raw_item.merge("retry_count" => 2, "failed_at" => now_f) }
-
-      context "without job in retry" do
-        it { is_expected.to match_array([digest]) }
-      end
-
-      context "with job in retry" do
-        before { zadd("retry", Time.now.to_f.to_s, dump_json(item)) }
-
-        it { is_expected.to match_array([]) }
-      end
-    end
-
-    context "when digest exists in a queue" do
-      context "without enqueued job" do
-        it { is_expected.to match_array([digest]) }
-      end
-
-      context "with enqueued job" do
-        before { push_item(item) }
-
-        it { is_expected.to match_array([]) }
-      end
-    end
-  end
-
   describe ".call" do
     subject(:call) { described_class.call }
